Moving permission helpers to mysqlctlproto.
