+++ go/vt/mysqlctl/permissions.go
@@ -5,7 +5,7 @@
 package mysqlctl
 
 import (
-	"github.com/youtube/vitess/go/vt/mysqlctl/proto"
+	"github.com/youtube/vitess/go/vt/mysqlctl/mysqlctlproto"
 
 	tabletmanagerdatapb "github.com/youtube/vitess/go/vt/proto/tabletmanagerdata"
 )
@@ -20,7 +20,7 @@ func GetPermissions(mysqld MysqlDaemon) (*tabletmanagerdatapb.Permissions, error
 		return nil, err
 	}
 	for _, row := range qr.Rows {
-		permissions.UserPermissions = append(permissions.UserPermissions, proto.NewUserPermission(qr.Fields, row))
+		permissions.UserPermissions = append(permissions.UserPermissions, mysqlctlproto.NewUserPermission(qr.Fields, row))
 	}
 
 	// get Dbs
@@ -29,7 +29,7 @@ func GetPermissions(mysqld MysqlDaemon) (*tabletmanagerdatapb.Permissions, error
 		return nil, err
 	}
 	for _, row := range qr.Rows {
-		permissions.DbPermissions = append(permissions.DbPermissions, proto.NewDbPermission(qr.Fields, row))
+		permissions.DbPermissions = append(permissions.DbPermissions, mysqlctlproto.NewDbPermission(qr.Fields, row))
 	}
 
 	return permissions, nil
